rename logger to logging to avoid confusion with variables named logger
