+++ controller-utils/pkg/logging/implementation.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package logger
+package logging
 
 import (
 	"github.com/go-logr/zapr"
